UPDATE: Scaffold database_v4

+++ srs/backend/WebService/src/Domain/Entities/CategoryProduct.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities;
 
 public partial class CategoryProduct
 {
-    public long CateProdId { get; set; }
+    public short CateProdId { get; set; }
 
     public string CateProdName { get; set; } = null!;
